requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
Deprecated==1.2.13
distlib==0.3.4
Django==4.1.1
djangorestframework==3.14.0
docutils==0.19
filelock==3.7.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
gcloud==0.18.3
googleapis-common-protos==1.56.4
greenlet==1.1.3
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.3.1
jws==0.1.3
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kivymd==1.1.1
MarkupSafe==2.1.1
oauth2client==4.1.3
Pillow==9.3.0
platformdirs==2.5.2
protobuf==4.21.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pygame==2.1.2
Pygments==2.13.0
pyparsing==3.0.9
pypiwin32==223
Pyrebase==3.0.27
Pyrebase4==4.5.0
python-jwt==2.0.1
pytz==2022.6
pywin32==305
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.41
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.11
virtualenv==20.15.1
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
